requirements.txt
Django==1.10.4
django-braces==1.10.0
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
drf-extensions==0.3.1
oauthlib==1.1.2
six==1.10.0
